refactor(worktree): split lib.ts into focused lib/ modules
lib.ts was a 468-line grab-bag. Split into single-responsibility modules with
no import cycles (ports → exec → registry/git → supabase/migrate →
launch-env/services/deps). lib.ts is now a barrel re-exporting all of them, so
cli.ts and the tests import from './lib' unchanged — the public API is byte-for-
byte identical. Code moved verbatim (domain comments preserved), no behaviour
change.

contract.test.ts now scans the whole lib/ dir (not just lib.ts) for spawned
binaries + @kortix/db script refs, so the guard still covers the split source.

Verified: bun build resolves all modules, tsc --strict clean, 19 tests pass.
